[incidents][JS]Improved clear session storage data clicking in the save and continue button, closes #229

+++ static/js/admin_change_form.js
@@ -3,10 +3,13 @@ document.addEventListener("DOMContentLoaded", function () {
     const inputSelector = 'input:not([type="number"]):not([type="checkbox"]):not([type="hidden"]), textarea';
     const langTabsSelector = '.parler-language-tabs a';
     const saveButtonSelector = 'input[name="_save"]';
+    const addAnotherButtonSelector = 'input[name="_addanother"]';
+    const continueButtonSelector = 'input[name="_continue"]';
     const currentUrl = window.location.href.split('?')[0];
     const urlId = currentUrl.split("/admin/")[1];
     const langTabs = Array.from(document.querySelectorAll(langTabsSelector))
     .filter(tab => !tab.classList.contains('deletelink'));
+    let clear_session_storage = true;
 
     function saveUrlId() {
         checkUrlId()
@@ -16,7 +19,7 @@ document.addEventListener("DOMContentLoaded", function () {
     function checkUrlId() {
         const storedUrlId  = sessionStorage.getItem('urlId');
 
-        if (urlId && !window.location.href.includes(storedUrlId)) {
+        if (urlId && !window.location.href.includes(storedUrlId) && clear_session_storage) {
             sessionStorage.clear();
         }
     }
@@ -64,11 +67,22 @@ document.addEventListener("DOMContentLoaded", function () {
     if (currentTabElement) restoreTabData(currentTabElement.name);
 
     const saveButton = document.querySelector(saveButtonSelector);
+    const addAnotherButton = document.querySelector(addAnotherButtonSelector);
+    const continueButton = document.querySelector(continueButtonSelector);
+
     if (saveButton) {
         saveButton.addEventListener('click', function () {
             sessionStorage.clear();
         });
     }
-
-
+    if (addAnotherButton) {
+        addAnotherButton.addEventListener('click', function () {
+            sessionStorage.clear();
+        });
+    }
+    if (continueButton) {
+        continueButton.addEventListener('click', function () {
+            clear_session_storage= false;
+        });
+    }
 });
